requirements.txt
torch
torchvision
opencv-python
numpy
pillow
matplotlib
tensorboard
scikit-learn
seaborn